requirements.txt
mysql-connector-python>=8.0.0
Pillow>=9.0.0
openpyxl>=3.0.0
reportlab>=3.6.0
tkcalendar>=1.6.1
